Delete code & compress repo to necessary packages
This is the first step in compressing this repo down into a pair of
reusable packages. The goal is to create packages that will allow other
go projects to programmatically create junit reports and print them out
to a writer of their choice.

This is contrasted with the original purpose of this forked repo which
was to convert the console output of go tests into a junit report.

To get started the project has been simplified down to two packages
which perform the following functions:

1. junit: The junit package contains all the relevant junit types as
well as a function that allows clients to write out a junit test suite
to a given writer.

2. testrun: The testrun package contains basic report, package, and
test-related types that are intended to be used by a client to track the
results of their tests. It also contains a convenience function to
convert a report into a testsuite type.