Consolidate repo into a single package
Further consolidated the repository into a single junit package. Also
updated a number of structs and functions that did not need to be
exposed.

Moving forward the only things that will be exposed are the Report
structs in report.go along with the Write function in writer.go.
Everything else is pretty much implementation details and don't need to
be exposed.